Canary Deployment: Revert - The tests extensibility is useless.
